Moving st to deltagenerator
