+++ lib/streamlit/DeltaGenerator.py
@@ -22,21 +22,32 @@ from streamlit.compatibility import setup_2_3_shims
 setup_2_3_shims(globals())
 
 import functools
+import contextlib
+import traceback
+import threading
 import json
 import random
 import textwrap
+import re
 import pandas as pd
+import sys as _sys
 from datetime import datetime
 from datetime import date
 from datetime import time
 
 from streamlit import caching
 from streamlit import metrics
+from streamlit import util as _util
 from streamlit.proto import Balloons_pb2
 from streamlit.proto import BlockPath_pb2
 from streamlit.proto import ForwardMsg_pb2
 from streamlit.proto import Text_pb2
-from streamlit import get_report_ctx
+from streamlit import code_util as _code_util
+from streamlit.ReportThread import get_report_ctx, add_report_ctx
+from streamlit.compatibility import (
+    setup_2_3_shims as _setup_2_3_shims,
+    is_running_py3 as _is_running_py3,
+)
 
 # setup logging
 from streamlit.logger import get_logger
@@ -2150,6 +2161,349 @@ class DeltaGenerator(object):
         import streamlit.elements.data_frame_proto as data_frame_proto
 
         data_frame_proto.marshall_data_frame(data, element.table)
+
+    def show(self, *args):
+        """Write arguments to your app for debugging purposes.
+
+        Show() has similar properties to write():
+
+            1. You can pass in multiple arguments, all of which will be
+            debugged.
+            2. It returns None, so it's "slot" in the app cannot be reused.
+
+        Parameters
+        ----------
+        *args : any
+            One or many objects to debug in the App.
+
+        Example
+        -------
+
+        >>> dataframe = pd.DataFrame({
+        ...     'first column': [1, 2, 3, 4],
+        ...     'second column': [10, 20, 30, 40],
+        ... }))
+        >>> st.show(dataframe)
+
+        Notes
+        -----
+        This is an experimental feature with usage limitations.
+
+        - The method must be called with the name `show`
+        - Must be called in one line of code, and only once per line
+        - When passing multiple arguments the inclusion of `,` or `)` in a
+         string argument may cause an error.
+
+        """
+        if not args:
+            return
+
+        try:
+            import inspect
+
+            # Get the calling line of code
+            previous_frame = inspect.currentframe().f_back.f_back
+            lines = inspect.getframeinfo(previous_frame)[3]
+
+            if not lines:
+                self.warning("`show` not enabled in the shell")
+                return
+
+            # Parse arguments from the line
+            line = lines[0].split("show", 1)[1]
+            inputs = _code_util.get_method_args_from_code(args, line)
+
+            # Escape markdown and add deltas
+            for idx, input in enumerate(inputs):
+                escaped = _util.escape_markdown(input)
+
+                self.markdown("**%s**" % escaped)
+                self.write(args[idx])
+
+        except Exception:
+            _, exc, exc_tb = _sys.exc_info()
+            self.exception(exc, exc_tb)  # noqa: F821
+
+    def write(self, *args, **kwargs):
+        """Write arguments to the app.
+
+        This is the swiss-army knife of Streamlit commands. It does different
+        things depending on what you throw at it.
+
+        Unlike other Streamlit commands, write() has some unique properties:
+
+            1. You can pass in multiple arguments, all of which will be written.
+            2. Its behavior depends on the input types as follows.
+            3. It returns None, so it's "slot" in the App cannot be reused.
+
+        Parameters
+        ----------
+        *args : any
+            One or many objects to print to the App.
+
+            Arguments are handled as follows:
+
+                - write(string)     : Prints the formatted Markdown string.
+                - write(data_frame) : Displays the DataFrame as a table.
+                - write(error)      : Prints an exception specially.
+                - write(func)       : Displays information about a function.
+                - write(module)     : Displays information about the module.
+                - write(dict)       : Displays dict in an interactive widget.
+                - write(obj)        : The default is to print str(obj).
+                - write(mpl_fig)    : Displays a Matplotlib figure.
+                - write(altair)     : Displays an Altair chart.
+                - write(keras)      : Displays a Keras model.
+                - write(graphviz)   : Displays a Graphviz graph.
+                - write(plotly_fig) : Displays a Plotly figure.
+                - write(bokeh_fig)  : Displays a Bokeh figure.
+
+        unsafe_allow_html : bool
+            This is a keyword-only argument that defaults to False.
+
+            By default, any HTML tags found in strings will be escaped and
+            therefore treated as pure text. This behavior may be turned off by
+            setting this argument to True.
+
+            That said, *we strongly advise* against it*. It is hard to write secure
+            HTML, so by using this argument you may be compromising your users'
+            security. For more information, see:
+
+            https://github.com/streamlit/streamlit/issues/152
+
+            *Also note that `unsafe_allow_html` is a temporary measure and may be
+            removed from Streamlit at any time.*
+
+            If you decide to turn on HTML anyway, we ask you to please tell us your
+            exact use case here:
+
+            https://discuss.streamlit.io/t/96
+
+            This will help us come up with safe APIs that allow you to do what you
+            want.
+
+        Example
+        -------
+
+        Its simplest use case is to draw Markdown-formatted text, whenever the
+        input is a string:
+
+        >>> write('Hello, *World!*')
+
+        .. output::
+           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=DUJaq97ZQGiVAFi6YvnihF
+           height: 50px
+
+        As mentioned earlier, `st.write()` also accepts other data formats, such as
+        numbers, data frames, styled data frames, and assorted objects:
+
+        >>> st.write(1234)
+        >>> st.write(pd.DataFrame({
+        ...     'first column': [1, 2, 3, 4],
+        ...     'second column': [10, 20, 30, 40],
+        ... }))
+
+        .. output::
+           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=FCp9AMJHwHRsWSiqMgUZGD
+           height: 250px
+
+        Finally, you can pass in multiple arguments to do things like:
+
+        >>> st.write('1 + 1 = ', 2)
+        >>> st.write('Below is a DataFrame:', data_frame, 'Above is a dataframe.')
+
+        .. output::
+           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=DHkcU72sxYcGarkFbf4kK1
+           height: 300px
+
+        Oh, one more thing: `st.write` accepts chart objects too! For example:
+
+        >>> import pandas as pd
+        >>> import numpy as np
+        >>> import altair as alt
+        >>>
+        >>> df = pd.DataFrame(
+        ...     np.random.randn(200, 3),
+        ...     columns=['a', 'b', 'c'])
+        ...
+        >>> c = alt.Chart(df).mark_circle().encode(
+        ...     x='a', y='b', size='c', color='c')
+        >>>
+        >>> st.write(c)
+
+        .. output::
+           https://share.streamlit.io/0.25.0-2JkNY/index.html?id=8jmmXR8iKoZGV4kXaKGYV5
+           height: 200px
+
+        """
+        # Python2 doesn't support this syntax
+        #   def write(*args, unsafe_allow_html=False)
+        # so we do this instead:
+        unsafe_allow_html = kwargs.get("unsafe_allow_html", False)
+
+        from streamlit import _DATAFRAME_LIKE_TYPES, _HELP_TYPES
+
+        try:
+            string_buffer = []
+
+            def flush_buffer():
+                if string_buffer:
+                    self.markdown(
+                        " ".join(string_buffer),
+                        unsafe_allow_html=unsafe_allow_html
+                    )  # noqa: F821
+                    string_buffer[:] = []
+
+            for arg in args:
+                # Order matters!
+                if isinstance(arg, string_types):  # noqa: F821
+                    string_buffer.append(arg)
+                elif type(arg).__name__ in _DATAFRAME_LIKE_TYPES:
+                    flush_buffer()
+                    import numpy as _np
+                    if len(_np.shape(arg)) > 2:
+                        self.text(arg)
+                    else:
+                        self.dataframe(arg)  # noqa: F821
+                elif isinstance(arg, Exception):
+                    flush_buffer()
+                    self.exception(arg)  # noqa: F821
+                elif isinstance(arg, _HELP_TYPES):
+                    flush_buffer()
+                    self.help(arg)
+                elif _util.is_altair_chart(arg):
+                    flush_buffer()
+                    self.altair_chart(arg)
+                elif _util.is_type(arg, "matplotlib.figure.Figure"):
+                    flush_buffer()
+                    self.pyplot(arg)
+                elif _util.is_plotly_chart(arg):
+                    flush_buffer()
+                    self.plotly_chart(arg)
+                elif _util.is_type(arg, "bokeh.plotting.figure.Figure"):
+                    flush_buffer()
+                    self.bokeh_chart(arg)
+                elif _util.is_graphviz_chart(arg):
+                    flush_buffer()
+                    self.graphviz_chart(arg)
+                elif _util.is_keras_model(arg):
+                    from tensorflow.python.keras.utils import vis_utils
+
+                    flush_buffer()
+                    dot = vis_utils.model_to_dot(arg)
+                    self.graphviz_chart(dot.to_string())
+                elif (type(arg) in dict_types) or (
+                isinstance(arg, list)):  # noqa: F821
+                    flush_buffer()
+                    self.json(arg)
+                elif _util.is_namedtuple(arg):
+                    flush_buffer()
+                    self.json(json.dumps(arg._asdict()))
+                else:
+                    string_buffer.append("`%s`" % str(arg).replace("`", "\\`"))
+
+            flush_buffer()
+
+        except Exception:
+            _, exc, exc_tb = _sys.exc_info()
+            self.exception(exc, exc_tb)  # noqa: F821
+
+    @contextlib.contextmanager
+    def spinner(self, text="In progress..."):
+        """Temporarily displays a message while executing a block of code.
+
+        Parameters
+        ----------
+        text : str
+            A message to display while executing that block
+
+        Example
+        -------
+
+        >>> with st.spinner('Wait for it...'):
+        >>>     time.sleep(5)
+        >>> st.success('Done!')
+
+        """
+
+        import streamlit.caching as caching
+
+        display_message_lock = None
+
+        # @st.cache optionally uses spinner for long-running computations.
+        # Normally, streamlit warns the user when they call st functions
+        # from within an @st.cache'd function. But we do *not* want to show
+        # these warnings for spinner's message, so we create and mutate this
+        # message delta within the "suppress_cached_st_function_warning"
+        # context.
+        with caching.suppress_cached_st_function_warning():
+            message = self.empty()
+
+        try:
+            # Set the message 0.1 seconds in the future to avoid annoying
+            # flickering if this spinner runs too quickly.
+            DELAY_SECS = 0.1
+            display_message = True
+            display_message_lock = threading.Lock()
+
+            def set_message():
+                with display_message_lock:
+                    if display_message:
+                        with caching.suppress_cached_st_function_warning():
+                            message.warning(str(text))
+
+            add_report_ctx(threading.Timer(DELAY_SECS, set_message)).start()
+
+            # Yield control back to the context.
+            yield
+        finally:
+            if display_message_lock:
+                with display_message_lock:
+                    display_message = False
+            with caching.suppress_cached_st_function_warning():
+                message.empty()
+
+    @contextlib.contextmanager
+    def echo(self):
+        """Use in a `with` block to draw some code on the app, then execute it.
+
+        Example
+        -------
+
+        >>> with st.echo():
+        >>>     st.write('This code will be printed')
+
+        """
+
+        spaces_re = re.compile("\\s*")
+
+        code = self.empty()
+
+        try:
+            frame = traceback.extract_stack()[-3]
+            if _is_running_py3():
+                filename, start_line = frame.filename, frame.lineno
+            else:
+                filename, start_line = frame[:2]
+            yield
+            frame = traceback.extract_stack()[-3]
+            if _is_running_py3():
+                end_line = frame.lineno
+            else:
+                end_line = frame[1]
+            lines_to_display = []
+            with open(filename) as source_file:
+                source_lines = source_file.readlines()
+                lines_to_display.extend(source_lines[start_line:end_line])
+                initial_spaces = spaces_re.match(lines_to_display[0]).end()
+                for line in source_lines[end_line:]:
+                    if spaces_re.match(line).end() < initial_spaces:
+                        break
+                    lines_to_display.append(line)
+            lines_to_display = textwrap.dedent("".join(lines_to_display))
+            code.code(lines_to_display, "python")
+
+        except FileNotFoundError as err:  # noqa: F821
+            code.warning("Unable to display code. %s" % err)
 
     def add_rows(self, data=None, **kwargs):
         """Concatenate a dataframe to the bottom of the current one.
