+++ lib/tests/streamlit/keras_test.py
@@ -36,7 +36,7 @@ class KerasTest(unittest.TestCase):
         model.add(Flatten())
         model.add(Dense(8, activation="relu"))
 
-        with patch("streamlit.graphviz_chart") as graphviz_chart:
+        with patch("streamlit.DeltaGenerator.DeltaGenerator.graphviz_chart") as graphviz_chart:
             st.write(model)
 
             dot = vis_utils.model_to_dot(model)
