+++ lib/tests/streamlit/write_test.py
@@ -37,17 +37,21 @@ class StreamlitWriteTest(unittest.TestCase):
     trying to check is that the right st.* method gets called
     """
 
+    def _patch_delta_generator_method(self, method):
+        return patch("streamlit.DeltaGenerator.DeltaGenerator." + method)
+
     def test_string(self):
         """Test st.write with a string."""
-        with patch("streamlit.markdown") as p:
+        with self._patch_delta_generator_method("markdown") as p:
             st.write("some string")
 
             p.assert_called_once()
 
-        with patch("streamlit.markdown") as p:
+        with self._patch_delta_generator_method("markdown") as p:
             st.write("more", "strings", "to", "pass")
 
-            p.assert_called_once_with("more strings to pass", unsafe_allow_html=False)
+            p.assert_called_once_with("more strings to pass",
+                                      unsafe_allow_html=False)
 
     def test_dataframe(self):
         """Test st.write with dataframe."""
@@ -63,14 +67,14 @@ class StreamlitWriteTest(unittest.TestCase):
         self.assertEqual(sorted(data.keys()), sorted(st._DATAFRAME_LIKE_TYPES))
 
         for df in data.values():
-            with patch("streamlit.dataframe") as p:
+            with self._patch_delta_generator_method("dataframe") as p:
                 st.write(df)
 
                 p.assert_called_once()
 
     def test_exception_type(self):
         """Test st.write with exception."""
-        with patch("streamlit.exception") as p:
+        with self._patch_delta_generator_method("exception") as p:
             st.write(Exception("some exception"))
 
             p.assert_called_once()
@@ -78,13 +82,13 @@ class StreamlitWriteTest(unittest.TestCase):
     def test_help(self):
         """Test st.write with help types."""
         # Test module
-        with patch("streamlit.help") as p:
+        with self._patch_delta_generator_method("help") as p:
             st.write(np)
 
             p.assert_called_once()
 
         # Test function
-        with patch("streamlit.help") as p:
+        with self._patch_delta_generator_method("help") as p:
             st.write(st.set_option)
 
             p.assert_called_once()
@@ -97,7 +101,7 @@ class StreamlitWriteTest(unittest.TestCase):
         class FakeChart(object):
             pass
 
-        with patch("streamlit.altair_chart") as p:
+        with self._patch_delta_generator_method("altair_chart") as p:
             st.write(FakeChart())
 
             p.assert_called_once()
@@ -110,7 +114,7 @@ class StreamlitWriteTest(unittest.TestCase):
         class FakePyplot(object):
             pass
 
-        with patch("streamlit.pyplot") as p:
+        with self._patch_delta_generator_method("pyplot") as p:
             st.write(FakePyplot())
 
             p.assert_called_once()
@@ -119,36 +123,36 @@ class StreamlitWriteTest(unittest.TestCase):
         import plotly.graph_objs as go
 
         """Test st.write with plotly object."""
-        with patch("streamlit.plotly_chart") as p:
+        with self._patch_delta_generator_method("plotly_chart") as p:
             st.write([go.Scatter(x=[1, 2], y=[10, 20])])
 
             p.assert_called_once()
 
     def test_dict(self):
         """Test st.write with dict."""
-        with patch("streamlit.json") as p:
+        with self._patch_delta_generator_method("json") as p:
             st.write({"a": 1, "b": 2})
 
             p.assert_called_once()
 
     def test_list(self):
         """Test st.write with list."""
-        with patch("streamlit.json") as p:
+        with self._patch_delta_generator_method("json") as p:
             st.write([1, 2, 3])
 
             p.assert_called_once()
 
     def test_namedtuple(self):
         """Test st.write with list."""
-        with patch("streamlit.json") as p:
+        with self._patch_delta_generator_method("json") as p:
             Boy = namedtuple("Boy", ("name", "age"))
             John = Boy("John", 29)
             st.write(John)
 
             p.assert_called_once()
 
-    @patch("streamlit.markdown")
-    @patch("streamlit.json")
+    @patch("streamlit.DeltaGenerator.DeltaGenerator.markdown")
+    @patch("streamlit.DeltaGenerator.DeltaGenerator.json")
     def test_dict_and_string(self, mock_json, mock_markdown):
         """Test st.write with dict."""
         manager = Mock()
@@ -171,7 +175,7 @@ class StreamlitWriteTest(unittest.TestCase):
             def __str__(self):
                 return "1 * 2 - 3 = 4 `ok` !"
 
-        with patch("streamlit.markdown") as p:
+        with self._patch_delta_generator_method("markdown") as p:
             st.write(SomeObject())
 
             p.assert_called_once_with(
@@ -180,7 +184,7 @@ class StreamlitWriteTest(unittest.TestCase):
 
     def test_exception(self):
         """Test st.write that raises an exception."""
-        with patch("streamlit.markdown") as m, patch("streamlit.exception") as e:
+        with self._patch_delta_generator_method("markdown") as m, self._patch_delta_generator_method("exception") as e:
             m.side_effect = Exception("some exception")
             st.write("some text")
 
@@ -190,7 +194,7 @@ class StreamlitWriteTest(unittest.TestCase):
         """Test st.spinner."""
         # TODO(armando): Test that the message is actually passed to
         # message.warning
-        with patch("streamlit.empty") as e:
+        with self._patch_delta_generator_method("empty") as e:
             with st.spinner("some message"):
                 time.sleep(0.15)
             e.assert_called_once_with()
